07_dashboard/backend/requirements.txt
anyio==4.9.0
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
croniter==6.0.0
exceptiongroup==1.3.0
Flask==3.1.1
flask-cors==6.0.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
notion-client==2.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.4
six==1.17.0
sniffio==1.3.1
typing_extensions==4.14.0
urllib3==2.4.0
Werkzeug==3.1.3
